% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_vec.R
\name{dist_vec}
\alias{dist_vec}
\title{compute distances between two vectors of points}
\usage{
dist_vec(idxy1, idxy2, row_id = T)
}
\arguments{
\item{idxy1}{first vector of coordinates}

\item{idxy2}{second vector of coordinates}

\item{row_id}{should row ids be returned}
}
\value{
<Delete and Replace>
}
\description{
compute distances between two vectors of points
}
\details{
This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 12/7/2023 Created \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 dat1 = data.frame(id=1:25, x=25:49, y=61:85)
 dat2 = data.frame(id=LETTERS[1:10], x=rnorm(10), y=runif(10))
 dist_vec(dat1, dat2)


}
\seealso{
\code{\link{dist}}\cr \code{\link{sweep}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
