% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.r
\name{.estimate}
\alias{.estimate}
\title{This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.}
\usage{
.estimate(
  x,
  resp_nm = "volcfgrs",
  aux_nm = "ht_p70",
  strata_nm = "str",
  wt_nm = c(NA, "wt"),
  ef_nm = c(NA, "ef"),
  su_nm = c(NA, "su"),
  reg_form = "ht_p20 + ht_p70 * percentage_first_returns_above_6_00",
  pop,
  N,
  strata_cuts,
  type = c("ht", "regression", "stratified", "two-stage", "calibrate"),
  var_type = c("asym", "bs", "svypkg")
)
}
\description{
This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.
}
