% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated2.R
\name{duplicated2}
\alias{duplicated2}
\title{same as duplicated but argument to show all duplicates}
\usage{
duplicated2(..., all = T)
}
\arguments{
\item{...}{regular arguments to duplicated}

\item{all}{T/F return all duplicates}
}
\value{
logical vector flagging duplicated elements
}
\description{
same as duplicated but argument to show all duplicates
}
\details{
same as duplicated but argument to show all duplicates

\cr
Revision History
\tabular{ll}{
1.0 \tab 2023 May 12 Created\cr
}
}
\examples{
 x <- c("a", "b", "a", "c", "b")

 # base duplicated() flags only the 2nd+ occurrence
 duplicated2(x, all = FALSE)

 # all = TRUE flags every element that has a duplicate
 duplicated2(x, all = TRUE)

}
\seealso{
\code{\link{duplicated}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
