% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stamp.R
\name{file_stamp}
\alias{file_stamp}
\title{create a date based version of a file name, mostly a wrapper for timestamp}
\usage{
file_stamp(
  path = tempdir(),
  stamp = format(Sys.time(), "\%Y\%m\%d\%H\%M\%S"),
  prefix = "logfile_V",
  suffix = ".txt",
  quiet = T
)
}
\arguments{
\item{path}{directory for the file, defaults to tempdir()}

\item{stamp}{how to label file - default is a custom format time}

\item{prefix}{some sort of label like function name or analysis}

\item{suffix}{file extension or other details}

\item{quiet}{If TRUE, suppress printing file_stamp to the console}
}
\value{
a character string: \code{file.path(path, paste0(prefix, stamp, suffix))}
}
\description{
create a date based version of a file name leveraging timestamp function and data format codes
}
\details{
create a date based version of a file name file name leveraging existing timestamp function and data format codes
#'

 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab Implemented  \cr
}
}
\examples{

#defaults: a timestamped file name under tempdir()
file_stamp()

#supply the stamp explicitly (note: stamp= is the SECOND argument; the first
#argument is the directory `path`)
file_stamp(stamp = format(Sys.time(), "\%Y_\%m_\%d_\%H\%M\%S"))
file_stamp(stamp = "1.1")
file_stamp(path = "c:/temp/project1", stamp = "V1.1", prefix = "myfile_", suffix = ".RDS")


}
\seealso{
\code{\link[utils]{timestamp}}\cr \code{\link{format}}\cr  \code{\link{Sys.time}}\cr \code{\link{tempfile}}\cr
}
\author{
Some Body <some.body@somewhere.com>
}
