% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvs_prototype_params.R
\name{fvs_prototype_params}
\alias{fvs_prototype_params}
\title{make template parameter data.frame}
\usage{
fvs_prototype_params(add_params = NULL)
}
\arguments{
\item{add_params}{optional substitution parameters, but must match substitution in keyfile using @somekey@}
}
\value{
data.frame with parameters for FVS key files
}
\description{
build an empty (0-row) template data.frame of the FVS parameter columns
expected by \code{fvs_make_keyfiles}/\code{fvs_run}
}
\details{
returns a 0-row data.frame whose columns are the standard FVS run parameters
(\code{std_cn}, \code{std_id}, \code{invyr}, ...); \code{add_params} appends
extra column names for any custom \code{@token@} substitutions

\cr
Revision History
\tabular{ll}{
1.0 \tab 2021 Oct 15 Created\cr
}
}
\examples{

#empty template
df_params = fvs_prototype_params()
str(df_params)

\dontrun{
#assume a typical inventory dataset and prepare fvs parameters
stand_data = data.frame(stand=1:10, year=2000:2009)
df_params = fvs_prototype_params()
df_params[1:nrow(stand_data),]=NA
df_params[,"std_id"] = stand_data$stand
df_params[,"invyr"] = stand_data$year
df_params[,"timeint"] = 1
df_params[,"numcycle"] = 1
df_params[,"input_db"] = "c:/temp/fordata.db"
df_params[,"fvs_path"] = "C:/FVSbin/FVSca.exe"
df_params[,"tree_table"] = "fvs_treeinit"
df_params[,"stand_table"] = "fvs_standinit"

#prepare prototype key file
key_proto = fvs_prototype_keyfile(invyr = "InvYear       2001", notriple=NULL)

#convert prototype key file into series of key files associated with each stand
clus1 = parallel::makeCluster(2)
df_keys = fvs_make_keyfiles(df_params, key_proto = key_proto, cluster = clus1, id="std_id")
fvs_run(df_keys, cluster = clus1)
parallel::stopCluster(clus1)
}


}
\seealso{
\code{\link{fvs_run}}\cr \code{\link{fvs_prototype_keyfile}}\cr \code{\link{fvs_make_keyfiles}}\cr  \code{\link{fvs_load}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
