% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvs_run.R
\name{fvs_run}
\alias{fvs_run}
\title{run fvs after preparing key files}
\usage{
fvs_run(
  key_df,
  db_merge = "FVSOut.db",
  fvs_commands = "FVS_commands.txt",
  merge_dbs = T,
  cluster = NA,
  clear_db = T,
  delete_temp_db = T,
  append = F,
  skip_empty = T
)
}
\arguments{
\item{key_df}{input dataframe with key file parameters}

\item{db_merge}{merge output parallel databases}

\item{fvs_commands}{optional file name to write the FVS command lines to (NA to skip)}

\item{merge_dbs}{T/F should temp dbs be merged into a single databse}

\item{cluster}{optional parallel cluster object}

\item{clear_db}{T/F delete all old inputs}

\item{delete_temp_db}{T/F delete temp db's used for parallel processing}

\item{append}{T/F if the output db already exists, should it be appended or wiped}

\item{skip_empty}{T/F skip empty per-run output databases when merging (otherwise error)}
}
\value{
invisibly, a list of the captured \code{system()} outputs for each FVS run
}
\description{
run fvs after preparing key files
1. prepare parameters
2. prepare prototype key file
3. generate key files for all cns
4. run fvs with run_fvs()
}
\details{
run fvs after preparing key files
1. prepare parameters
2. prepare prototype key file
3. generate key files for all cns
4. run fvs with run_fvs()

\cr
Revision History
\tabular{ll}{
1.0 \tab 2021 Oct 08 Created\cr
}
}
\examples{
\dontrun{
# requires a local FVS executable and input database
clus1 = parallel::makeCluster(2)

#assume a typical inventory dataset and prepare fvs parameters
stand_data = data.frame(stand=1:10, year=2000:2009)
df_params = fvs_prototype_params()
df_params[1:nrow(stand_data),]=NA
df_params[,"std_id"] = stand_data$stand
df_params[,"invyr"] = stand_data$year
df_params[,"timeint"] = 1
df_params[,"numcycle"] = 1
df_params[,"input_db"] = "c:/temp/fordata.db"
df_params[,"fvs_path"] = "C:/FVSbin/FVSca.exe"
df_params[,"tree_table"] = "fvs_treeinit"
df_params[,"stand_table"] = "fvs_standinit"

#prepare prototype key file
key_proto = fvs_prototype_keyfile(invyr = "InvYear       2001", notriple=NULL)

#convert prototype key file into series of key files associated with each stand
df_keys = fvs_make_keyfiles(df_params, key_proto = key_proto, cluster = clus1, id="std_id")

#lastly, actually run fvs
fvs_run(df_keys, cluster = clus1, db_merge = "FVS_AllRunsB.db")

parallel::stopCluster(clus1)
}

}
\seealso{
\code{\link{fvs_load}}\cr \code{\link{fvs_make_keyfiles}}\cr \code{\link{fvs_prototype_keyfile}}\cr \code{\link{fvs_prototype_params}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
