% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_functions}
\alias{helper_functions}
\alias{write_clipboard}
\alias{bs}
\alias{fs}
\alias{nms_vec}
\title{help with path slashes and data.frame name vectors}
\usage{
write_clipboard(x)

bs(x = NA)

fs(x = NA)

nms_vec(x)
}
\arguments{
\item{x}{(optional) path string}
}
\value{
either an R compatible path

or, in the case of nms_vec, a vector of R compatible data.frame column names

These clipboard helpers (\code{bs}, \code{fs}, \code{write_clipboard}) are
Windows-only: they call \code{utils::readClipboard}/\code{writeClipboard},
which exist only in Windows builds of R. On other platforms \code{bs()}/
\code{fs()} error unless you pass \code{x} explicitly, and
\code{write_clipboard} warns and does nothing. \code{nms_vec} works
everywhere (it just can't copy to the clipboard off Windows).
}
\description{
helper functions - path slashes and data.frame name vectors
}
\details{
helper functions - path slashes and data.frame name vectors

\cr
Revision History
\tabular{ll}{
1.0 \tab 2021 Oct 28 Created\cr
1.0 \tab 2024 Oct 01 add nms_vec example, move to underscores from camel calse\cr
}
}
\examples{

#nms_vec returns a quoted column-name vector (and copies it on Windows)
dftest = as.data.frame(t(1:30))
nms_vec(dftest)

\dontrun{
#these take a path on the clipboard (ctrl+C) and make it R compatible (Windows)
bs()
fs()
}

}
\seealso{
\code{\link{shQuote}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
