% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_tools.R
\name{knn_tools}
\alias{knn_tools}
\alias{yai_id}
\alias{newtargets_id}
\alias{impute_id}
\alias{yai_weights}
\alias{tl_impute}
\alias{tl_impute_2}
\alias{yai_cv}
\title{helpers for tree list imputation, knn}
\usage{
yai_id(
  xNms = NULL,
  yNms = NULL,
  idNm = NULL,
  data = NULL,
  omity = NULL,
  dup_ids_remove = T,
  ...
)

newtargets_id(yai_mod, idNm, data, k = NULL, ann = NULL)

impute_id(newtargs_id, ...)

yai_weights(
  yaimod,
  dtype = c("invdist", "invdist2", "eq"),
  zero_dist = c("small", "min", "NA")
)

tl_impute(
  wts,
  idNm,
  cols_knn_id = "col_id",
  cols_knn_wt = "col_wt",
  trees,
  sort_targets = F,
  debug = F
)

tl_impute_2(
  wts,
  idNm,
  cols_knn_id = "col_id",
  cols_knn_wt = "col_wt",
  trees,
  debug = F
)

yai_cv(
  omit = 5,
  idNm,
  xNm,
  yNm,
  pdNm = NA,
  data,
  iter_max = 500,
  k = 5,
  debug = FALSE,
  method = "msn",
  min_rows = 15,
  method_impute = c("closest", "mean", "median", "dstWeighted"),
  ...
)
}
\arguments{
\item{xNms}{character vector of auxiliary (predictor) column names}

\item{yNms}{character vector of response column names}

\item{idNm}{name of the id column (used as row names internally)}

\item{data}{data to use for knn}

\item{omity}{optional vector of ROW indices to omit (e.g. for cross-validation)}

\item{dup_ids_remove}{T/F drop records with duplicated ids (which break yaImpute)}

\item{...}{additional arguments to \code{yaImpute::yai} / downstream functions

\cr\cr
\bold{newtargets_id() parameters:}}

\item{yai_mod}{model returned by yai_id}

\item{k}{number of neighbors}

\item{ann}{T/F see yai documentation

\cr\cr
\bold{impute_id() parameters:}}

\item{newtargs_id}{object returned by newtargets_id()

\cr\cr
\bold{yai_weights() parameters:}}

\item{yaimod}{model returned by yai_id / newtargets_id}

\item{dtype}{distance-to-weight type: "invdist", "invdist2", or "eq"}

\item{zero_dist}{how to handle zero distances: "small", "min", or "NA"

\cr\cr
\bold{tl_impute() / tl_impute_2() parameters:}}

\item{wts}{weights structure returned by yai_weights()}

\item{cols_knn_id}{(reserved) name of the id-columns element in wts}

\item{cols_knn_wt}{(reserved) name of the weight-columns element in wts}

\item{trees}{tree-list records keyed by idNm}

\item{sort_targets}{T/F sort the output by target id}

\item{debug}{T/F drop into browser()

\cr\cr
\bold{yai_cv() parameters:}}

\item{omit}{number of observations to omit per cross-validation iteration}

\item{xNm}{character vector of auxiliary (predictor) column names}

\item{yNm}{character vector of response column names}

\item{pdNm}{columns to impute/predict (defaults to yNm)}

\item{iter_max}{maximum number of cross-validation iterations}

\item{method}{yai distance/imputation method (e.g. "msn","euclidean","rf")}

\item{min_rows}{minimum number of rows required to run cross-validation}

\item{method_impute}{imputation method passed to \code{yaImpute::impute}}
}
\value{
yai_id: \cr
newtargets_id: \cr
impute_id: \cr
yai_weights: \cr
tl_impute: \cr
yai_cv: \cr
}
\description{
Series of helpers which in combination with the package yaImpute can be used to
perform tree-list imputation.
}
\details{
yaImpute has some funkiness (e.g using row names as an ID field... bad...) and cannot support
tree list imputation. The functions here facilitate that process

yai_id: Update of yai that replaces row names with record ids
newtargets_id: wrapper for newtargets that handles yai_id object
impute_id: impute response values to target locations - wrapper for 'impute()' in yaImpute handles id field more explicitly
yai_weights: Get weights associated with imputation objects
tl_impute: function to impute tree list based on knn model fitted between aerial attributes
yai_cv: peform crossvalidation on knn model




 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.



\cr

Revision History
\tabular{ll}{
1.0 \tab  2014 July 04 Import from Jacob's R Library\cr
1.0 \tab  2015  Jan 06 Add roxygen header \cr
}
}
\examples{
\donttest{
 #simulated data: an id column, two auxiliary (x) and two response (y) variables
 set.seed(1)
 n <- 50
 dat <- data.frame(
   id = 1:n,
   x1 = rnorm(n), x2 = rnorm(n),
   y1 = rnorm(n), y2 = rnorm(n)
 )
 ref <- dat[1:40, ]    #reference (training) plots
 tgt <- dat[41:50, ]   #new target plots

 #fit a knn imputation model keyed on the id column (euclidean needs no extra deps)
 mod <- yai_id(
   xNms = c("x1", "x2"),
   yNms = c("y1", "y2"),
   idNm = "id",
   data = ref,
   method = "euclidean",
   k = 3
 )

 #impute to the new targets and pull the neighbour weights
 newt <- newtargets_id(mod, idNm = "id", data = tgt, k = 3, ann = FALSE)
 wts <- yai_weights(newt, dtype = "invdist")
 head(wts$wts)
}

}
\seealso{
\code{\link[yaImpute]{yai}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@usgda.gov>

\bold{yai_id() parameters:}
}
