% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVEL_biomass.R
\name{NVEL_biomass}
\alias{NVEL_biomass}
\title{Predict biomass using National Volume Estimator Library (NVEL)}
\usage{
NVEL_biomass(
  dfTL = data.frame(dbh = 15, ht = 65, cl = 30),
  bioeq = NA,
  geosub = NA,
  bioeqNm = c("bioeq"),
  geosubNm = "geosub",
  spcdNm = "spcd",
  dbhNm = "dbh",
  htNm = "ht",
  stemsNm = "stems",
  clNm = "cl",
  merchDbNm = c("merchDb"),
  htPrd1Nm = c("htPrd1"),
  htPrd2Nm = c("htPrd2"),
  cv4Nm = "cv4",
  dll_64 = system.file("lib/VolLibDll20231106/vollib-64bits/vollib.dll", package =
    "RForInvt"),
  dll_32 = system.file("lib/VolLibDll20231106/vollib-32bits/vollib.dll", package =
    "RForInvt"),
  load_dll = T,
  dll_func_bio = "biomasslib_r"
)
}
\arguments{
\item{dfTL}{tree list with a minimum of diamter and height columns}

\item{bioeq}{(optional) use single equation for all trees}

\item{bioeqNm}{(optional) column name in DFTL: provide region, forest, district for every tree in dfTL}

\item{spcdNm}{(required) column name in DFTL: USFS species code}

\item{dbhNm}{(required) column name in DFTL: dbh in inches}

\item{htNm}{(required) column name in DFTL: height in feet}

\item{clNm}{(optional) column name in DFTL: crown length in feet}

\item{merchDbNm}{(optional) column name in DFTL: upper db miminum for saw (in)}

\item{htPrd1Nm}{(optional) column name in DFTL: Height to the minimum top diameter inside bark for primary product (ft)}

\item{htPrd2Nm}{(optional) column name in DFTL: Height to the minimum top diameter inside bark for secondary product (ft)}

\item{dll_64}{path to 64bit dll}

\item{dll_32}{path to 64bit dll}

\item{load_dll}{T/F should dll be loaded (in case it is already loaded)}

\item{dll_func_vol}{name of volume prediction function call in NVEL .dll}

\item{dll_func_bioeq}{name of volume equation chooser function call in NVEL .dll}

\item{nclus}{number of cores to use}
}
\value{
a data.frame reformatted to include all of the original columns, any missing columns for the NVEL .dll, 15 columns of predicted volumes, and an error column
}
\description{
Predict biomass using National Volume Estimator Library (NVEL)
}
\details{
Predict biomass using National Volume Estimator Library (NVEL)

 see:
 https://www.fs.fed.us/forestmanagement/products/measurement/volume/nvel/index.php

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/19/2020 Implemented \cr
}
}
\examples{
\donttest{
        #a small inline tree list (USFS species codes, dbh in inches, ht in feet)
        df_fake <- data.frame(
            trid     = 1:4,
            region   = 6,
            forest   = "01",
            district = "01",
            dbh      = c(8, 12, 18, 24),
            ht       = c(45, 70, 95, 110),
            cl       = c(20, 35, 45, 55),
            spcd     = c(202, 122, 242, 263)
        )

        #predict green / dry biomass (requires the bundled vollib.dll)
        NVEL_biomass(dfTL = df_fake)
}


}
\author{
Jacob Strunk <someone@somewhere.com>
}
