% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVEL_voleq.R
\name{NVEL_voleq}
\alias{NVEL_voleq}
\title{Select Appropriate Volume equation for Region, Forest, District, Tree species from National Volume Estimator Library (NVEL)}
\usage{
NVEL_voleq(
  dfTL = NA,
  region = NA,
  forest = NA,
  district = NA,
  spcd = NA,
  regionNm = "region",
  forestNm = "forest",
  districtNm = "district",
  spcdNm = "spcd",
  dll_64 = system.file("lib/VolLibDll20231106/vollib-64bits/vollib.dll", package =
    "RForInvt"),
  dll_32 = system.file("lib/VolLibDll20231106/vollib-32bits/vollib.dll", package =
    "RForInvt"),
  dll_func_voleq = "getvoleq_r",
  load_dll = T
)
}
\arguments{
\item{dfTL}{data.frame with tree records}

\item{region}{scalar region; used only when dfTL is not supplied (otherwise the regionNm column is used)}

\item{forest}{scalar forest; used only when dfTL is not supplied (otherwise the forestNm column is used)}

\item{district}{scalar district; used only when dfTL is not supplied (otherwise the districtNm column is used)}

\item{spcd}{scalar USFS species code; used only when dfTL is not supplied (otherwise the spcdNm column is used)}

\item{regionNm}{(optional) column name in dfTL holding the region for every tree}

\item{forestNm}{(optional) column name in dfTL holding the forest for every tree}

\item{districtNm}{(optional) column name in dfTL holding the district for every tree}

\item{spcdNm}{(required) column name in dfTL: USFS species code}

\item{dll_64}{path to 64bit dll}

\item{dll_32}{path to 32bit dll}

\item{dll_func_voleq}{name of volume equation chooser function call in NVEL .dll}

\item{load_dll}{T/F should dll be loaded (in case it is already loaded)}
}
\value{
reformatted tree list with with a new column volume equation codes for NVEL - "voleq"
}
\description{
Select Appropriate Volume equation for Region, Forest, District, Tree species from National Volume Estimator Library (NVEL)
}
\details{
Select Appropriate Volume equation for Region, Forest, District, Tree species from National Volume Estimator Library (NVEL)

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/19/2020 Function created \cr
}
}
\examples{
\donttest{
        #look up volume equations (requires the bundled vollib.dll)
        NVEL_voleq(region = 2, forest = "01", district = "01", spcd = 951)
        NVEL_voleq(region = 2, forest = "01", district = "01", spcd = rep(c(951, 201), 2))
        NVEL_voleq(dfTL = data.frame(region = 6, forest = "01", district = "01",
                                     spcd = rep(c(951, 201), 2)))

        #a small inline tree list (USFS species codes)
        df_fake <- data.frame(
            trid     = 1:4,
            region   = 6,
            forest   = "01",
            district = "01",
            dbh      = c(8, 12, 18, 24),
            ht       = c(45, 70, 95, 110),
            spcd     = c(202, 122, 242, 263)
        )

        #choose a volume equation for every tree
        NVEL_voleq(dfTL = df_fake)
}

}
\seealso{
\code{\link{NVEL_volume}}\cr \code{\link{NVEL_biomass}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
