% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVEL_volume.R
\name{NVEL_volume}
\alias{NVEL_volume}
\title{Predict volume using National Volume Estimator Library (NVEL)}
\usage{
NVEL_volume(
  dfTL = data.frame(dbh = 5, ht = 5),
  voleq = NA,
  region = 1,
  forest = NA,
  district = NA,
  voleqNm = c("voleq"),
  regionNm = c("region"),
  forestNm = c("forest"),
  districtNm = c("district"),
  spcdNm = "spcd",
  dbhNm = "dbh",
  htNm = "ht",
  pulpDbNm = c("pulpDb"),
  sawDbNm = c("sawDb"),
  htPrd1Nm = c("htPrd1"),
  htPrd2Nm = c("htPrd2"),
  upHt1Nm = c("upHt1"),
  upDb1Nm = c("upDb1"),
  stumpHtNm = "stumpHt",
  fclassNm = c("fclass"),
  dbtbhNm = c("dbtbh"),
  btrNm = c("btr"),
  vol2biomass = T,
  dll_64 = system.file("lib/VolLibDll20231106/vollib-64bits/vollib.dll", package =
    "RForInvt"),
  dll_32 = system.file("lib/VolLibDll20231106/vollib-32bits/vollib.dll", package =
    "RForInvt"),
  load_dll = T,
  dll_func_vol = "vollib_r",
  dll_func_voleq = "getvoleq_r"
)
}
\arguments{
\item{dfTL}{tree list with a minimum of diamter and height columns}

\item{voleq}{force a specific volume equation for all trees}

\item{region}{(optional) region,forest,district but these supercede values in dfTL columns regionNm,forestNm, districtNm}

\item{forest}{(optional) region,forest,district but these supercede values in dfTL columns regionNm,forestNm, districtNm}

\item{district}{(optional) region,forest,district but these supercede values in dfTL columns regionNm,forestNm, districtNm}

\item{voleqNm}{(optional) column name in DFTL: provide region, forest, district for every tree in dfTL}

\item{regionNm}{(optional) column name in DFTL:provide region, forest, district for every tree in dfTL}

\item{forestNm}{(optional) column name in DFTL:provide region, forest, district for every tree in dfTL}

\item{districtNm}{(optional) column name in DFTL: provide region, forest, district for every tree in dfTL}

\item{spcdNm}{(required) column name in DFTL: USFS species code}

\item{dbhNm}{(required) column name in DFTL: dbh in inches}

\item{htNm}{(required) column name in DFTL: height in feet}

\item{pulpDbNm}{(optional) column name in DFTL: upper db miminum for pulp (in)}

\item{sawDbNm}{(optional) column name in DFTL: upper db miminum for saw (in)}

\item{htPrd1Nm}{(optional) column name in DFTL: Height to the minimum top diameter inside bark for primary product (ft)}

\item{htPrd2Nm}{(optional) column name in DFTL: Height to the minimum top diameter inside bark for secondary product (ft)}

\item{upHt1Nm}{(optional) column name in DFTL: Upper stem height in feet where upper stem diameter was measured or where AVG1 is to be applied. OR For Region 8 the upper stem reference height defined by their profile model (ht0, ht4, ht7, ht9). OR For region 9 the UPSHT1 is the height to 7.6/9.6 top diameter when HT1PRD is not the height to 7.6/9.6 top diameter.}

\item{upDb1Nm}{(optional) column name in DFTL: Upper stem diameter measured at}

\item{stumpHtNm}{(optional) column name in DFTL: stump height (ft)}

\item{fclassNm}{(optional) column name in DFTL: Girard’s form class.  Diameter at the top of the first log given as a percent of DBH.}

\item{dbtbhNm}{(optional) column name in DFTL: Double bark thickness at breast height in inches}

\item{btrNm}{(optional) column name in DFTL: Bark thickness ratio given as the percent of diameter inside bark to diameter outside bark.  (dib/dob *100).}

\item{vol2biomass}{T/F Compute total drye and green biomass in lbs using weight factors and append weight factors (lbs / cubic foot)}

\item{dll_64}{path to 64bit dll}

\item{dll_32}{path to 64bit dll}

\item{load_dll}{T/F should dll be loaded (in case it is already loaded)}

\item{dll_func_vol}{name of volume prediction function call in NVEL .dll}

\item{dll_func_voleq}{name of volume equation chooser function call in NVEL .dll}

\item{nclus}{number of cores to use}
}
\value{
a data.frame reformatted to include all of the original columns, any missing columns for the NVEL .dll, 15 columns of predicted volumes, and an error column

 predicted volumes:
    1)	Total Cubic Volume from ground to tip
    2)	Gross Scribner board foot volume. Note: The VOL(2) is International ¼ board foot volume for Region 8 Forest 8, 9, 10, and 12 (except Andrew Pickens district); and Region 9 Forest 4,5,8,11,12,14,19,20,21,22,24, and 30 when using Clark profile equation.
    3)	Net Scribner board foot volume.
    4)	Gross merchantable cubic foot volume
    5)	Net merchantable cubic foot volume
    6)	Merchantable cordwood volume
    7)	Gross secondary product volume in cubic feet
    8)	Net secondary product volume in cubic feet
    9)	Secondary product in cordwood
    10)	Gross International ¼ board foot volume
    11)	Net International ¼ board foot volume
    12)	Gross secondary product in Scribner board feet
    13)	Net secondary product in Scribner board feet
    14)	Stump volume
    15)	Tip volume
}
\description{
Predict volume using National Volume Estimator Library (NVEL)
}
\details{
Predict volume using National Volume Estimator Library (NVEL)

 see:
 https://www.fs.fed.us/forestmanagement/products/measurement/volume/nvel/index.php

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/19/2020 Implemented \cr
1.1 \tab 5/21/2020 Add biomass and weight factors \cr
1.2 \tab 12/05/2023 Add lates NVEL and fix paths to dlls \cr
}
}
\examples{

        #look up volume equations
         library(RForInvt)
         NVEL_voleq(region = 2, forest = "01",district = "01", spcd=951)
         NVEL_voleq(region = 2, forest = "01",district = "01", spcd=951)
         NVEL_voleq(region = 2, forest = "01",district = "01", spcd=rep(c(951,201),2))
         NVEL_voleq(dfTL=data.frame(region = 6, forest = "01",district = "01", spcd=rep(c(951,201),2)))

        #grab list of species
        if(!"dfSpp" \%in\% ls()){
          library(RSQLite)
          db0 = dbConnect(RSQLite::SQLite(), system.file("misc/NBEL/BiomassEqns.db", package="RForInvt"))
          dfSpp = dbGetQuery(db0, paste("select * from tblspp"))
          dfCoeff = dbGetQuery(db0, paste("select * from BM_EQCoefs"))
          dbDisconnect(db0)
        }

        #build a fake tree list
        if("df_fake" \%in\% ls()){
          set.seed=111
          nfake=length(unique(dfCoeff$species_code))

          df_fake = data.frame(
            trid=1:(nfake)
            ,region = 6
            ,forest = "01"
            ,district = "01"
            ,dbh=10*abs(rnorm(nfake))
            ,ht=100*abs(rnorm(nfake))
            ,spcd = unique(dfCoeff$species_code)#'     sample(c("a","b","c","d") , nfake , T)
          )

        }

        #get volumes
        NVEL_volume( dfTL = df_fake )



}
\seealso{
\code{\link{NVEL_voleq}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
