% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandermod.R
\name{pandermod}
\alias{pandermod}
\title{update on the pander functions to do what I want to lm and simex objects including lists}
\usage{
pandermod(x, ...)
}
\arguments{
\item{x}{inpute model or list of models (lm objects)}

\item{df0}{optional dataframe for validation (out of box validation)}

\item{return_df}{T/F should the function return a data.frame}
}
\value{
data.frame or pander table
}
\description{
simple output, and accepts a list of lm objects
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 Aug 22 \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{pander}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
