% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForInvt.R
\name{RForInvt}
\alias{RForInvt}
\title{Tools for forest inventory estimation}
\usage{
RForInvt()
}
\value{
depends on function
}
\description{
Tools for forest inventory estimation
}
\details{
Tools for forest inventory estimation
 currently in a preliminary stage, but some functionality present

 - compilePlots
 - compileTrees
 - ols_modeling
 - knn_tools
 - yai_r2
 - sampleSystematic
 - NVEL...() related functions
 - fvs..() related functions

Eventually

- estimate()

the estimate() functionality (currently use .estimate() )likely won't be any greater than the survey package
but it will be tailored to typical forest inventory scenarios.


\cr
Revision History
\tabular{ll}{
1.0 \tab 2021 10 18 Created \cr
}
}
\examples{
eventually...

}
\seealso{
\code{\link{lidR}}\cr \code{\link{RSForInvt}}\cr
}
\author{
Jacob Strunk <strunky@gmail.com>
}
