% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSystematic.R
\name{sampleSystematic}
\alias{sampleSystematic}
\title{take (a) systematic sample(s) from a vector}
\usage{
sampleSystematic(N, n = NA, exact_n = T, as_idx = T)
}
\arguments{
\item{N}{vector length}

\item{n}{length of sample (optionally a vector of samples)}

\item{exact_n}{should n be matched exactly}

\item{as_idx}{should a numeric index be return (T) or T/F vector (F)}
}
\value{
vector of numeric indices or T/F vector of length N
 vector of 0's (not sampled) and group ids 1:n for sampled items if length(n) > 1
}
\description{
take (a) systematic sample(s) from a vector
}
\details{
This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 7/13/2021 Implemented \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{

 sampleSystematic(2000, 25)
 sampleSystematic(205, n=c(50,50))


}
\seealso{
\code{\link{sample}}\cr \code{\link{which}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
