% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_text.R
\name{search_text}
\alias{search_text}
\title{search for string in text files in a directory}
\usage{
search_text(
  match_string = "",
  directory = NA,
  file = NA,
  pattern = ".r",
  pattern_omit = c(NA, "dev_"),
  recursive = T,
  clus = NA
)
}
\arguments{
\item{match_string}{string to find within searched files}

\item{directory}{where to search files}

\item{file}{single file}

\item{pattern}{pattern to search for files}

\item{pattern_omit}{pattern to omit searched files}

\item{recursive}{look in subfolders?}

\item{clus}{number of cores, or a cluster object}
}
\value{
two column dataframe like:
 [1] file    matches
<0 rows> (or 0-length row.names)
}
\description{
search for string in text files in a directory
}
\details{
search for string in text files in a directory

\cr
Revision History
\tabular{ll}{
1.0 \tab 2015 June 19 Created \cr
}
}
\examples{
 #create a small temporary directory with a couple of text files to search
 tmp_dir <- file.path(tempdir(), "search_text_demo")
 dir.create(tmp_dir, showWarnings = FALSE)
 writeLines(c("x <- 1", "require(plyr)"),        file.path(tmp_dir, "a.R"))
 writeLines(c("y <- 2", "library(data.table)"),  file.path(tmp_dir, "b.R"))

 #find the .R files (and line numbers) containing "plyr"
 test <- search_text(directory = tmp_dir, pattern = "[.]R$", match_string = "plyr")
 print(test)

 #clean up
 unlink(tmp_dir, recursive = TRUE)

}
\seealso{
\code{\link{makeCluster}}\cr \code{\link{list.files}}\cr  \code{\link{readLines}}\cr \code{\link{grepl}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
