% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write_clipboard}
\alias{write_clipboard}
\title{help with path slashes and data.frame name vectors}
\usage{
write_clipboard(x)
}
\arguments{
\item{x}{(optional) path string}
}
\value{
either an R compatible path

or, in the case of nms_vec, a vector of R compatible data.frame column names
}
\description{
helper functions - path slashes and data.frame name vectors
}
\details{
helper functions - path slashes and data.frame name vectors

\cr
Revision History
\tabular{ll}{
1.0 \tab 2021 Oct 28 Created\cr
1.0 \tab 2024 Oct 01 add nms_vec example, move to underscores from camel calse\cr
}
}
\examples{

#these take a path on the clipboard (ctrl+C) and makes them R compatible
bs()
fs()
dftest=as.data.frame(t(1:30))
nms_vec(dftest)


}
\seealso{
\code{\link{readClipboard}}\cr \code{\link{writeClipboard}}\cr \code{\link{shQuote}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
