% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{writeClipboard2}
\alias{writeClipboard2}
\title{help with path slashes and data.frame name vectors}
\usage{
writeClipboard2(x)
}
\arguments{
\item{x}{(optional) path string}
}
\value{
NA
}
\description{
helper functions - path slashes and data.frame name vectors
}
\details{
helper functions - path slashes and data.frame name vectors

\cr
Revision History
\tabular{ll}{
1.0 \tab 2021 Oct 28 Created\cr
}
}
\examples{

#these take a path on the clipboard (ctrl+C) and makes them R compatible
bs()
fs()


}
\seealso{
\code{\link{readClipboard}}\cr \code{\link{writeClipboard}}\cr \code{\link{shQuote}}\cr
}
\author{
Jacob Strunk <strunky@gmail.com>
}
