% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated2.R
\name{duplicated2}
\alias{duplicated2}
\title{same as duplicated but argument to show all duplicates}
\usage{
duplicated2(..., all = T)
}
\arguments{
\item{...}{regular arguments to duplicated}

\item{all}{T/F return all duplicates}
}
\value{
vector of duplicates
}
\description{
same as duplicated but argument to show all duplicates
}
\details{
same as duplicated but argument to show all duplicates

\cr
Revision History
\tabular{ll}{
1.0 \tab 2023 May 12 Created\cr
}
}
\examples{
something

}
\seealso{
\code{\link{duplicated}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
