% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filestamp.R
\name{filestamp}
\alias{filestamp}
\title{create a date versioned file name, just a wrapper for timestamp}
\usage{
filestamp(
  stamp = format(Sys.time(), "\%Y\%m\%d\%H\%M\%S"),
  prefix = "c:/temp/somefile_V",
  suffix = ".csv",
  quiet = F
)
}
\arguments{
\item{stamp}{how to label file - default is a custom format time}

\item{prefix}{file path or other}

\item{suffix}{file extesion}

\item{quiet}{If TRUE, suppress printing filestamp to the console}
}
\value{
character string with some new d
}
\description{
create a date versioned file name leveragin timestamp function and data format codes
}
\details{
create a date versioned file name leveraging existing timestamp function and data format codes
#'

 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab Implemented  \cr
}
}
\examples{
 
filestamp()
filestamp(format(Sys.time(), "\%Y_\%m_\%d_\%H\%M\%S"))
filestamp("1.1")
filestamp("V1.1",prefix = "c:/temp/project1/myfile_",suffix=".RDS")


}
\seealso{
\code{\link{timestamp}}\cr \code{\link{format}}\cr  \code{\link{Sys.time}}\cr
}
\author{
Some Body <some.body@somewhere.com>
}
