% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_strata.R
\name{make_strata}
\alias{make_strata}
\alias{assign_strata}
\title{Make and assign strata based on two input columns}
\usage{
make_strata(
  data,
  x1 = "PlotStrata1",
  x2 = "Elev.P95",
  split_x1 = c("qt", "eq"),
  split_x2 = c("qt", "eq"),
  nest_x2 = T,
  n1 = 10,
  n2 = 10,
  min_recs = 7,
  precision = 0,
  collapse = T
)

assign_strata(strata, data)
}
\arguments{
\item{data}{input data}

\item{x1}{name of input x field}

\item{x2}{name of input y field}

\item{split_x1}{how to split - equal interval or quantile based}

\item{split_x2}{how to split - equal interval or quantile based}

\item{nest_x2}{should x2 be nested in x1, or should x1 and x2 be split independently}

\item{n1}{number of strata for x1}

\item{n2}{number of strata for x2}

\item{min_recs}{what is the minimum number of records in a stratum before collapse}

\item{precision}{what precision is retained for cutting numeric values into strata}
}
\value{
make_strata returns a data.frame of strata boundaries
 assign_strata takes a data.frame and appends a column with the stratum name for each record
}
\description{
Make and assign strata based on two input attributes and some parameters
 functions are make_strata and assign_strata
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2019 01 24 created \cr
1.1 \tab 2020 06 25 updated to allow more complex inputs \cr
}
}
\examples{



      #prepare pseudo data
      n=500
      hts = sample(1:150,n,T)
      dbhs = abs(hts/5 + rnorm(#' )*2)
      dat_test = data.frame(height = hts , dbh = dbhs, height_cat = cut(hts,10) , dbh_cat = cut(dbhs,10))
      plot(dbhs,hts)

      #example 1 factor and numeric
      str_test = make_strata(
        dat_test
        , x1="height_cat"
        , x2="dbh"
        , split_x1 =c("qt","eq")[1]
        , split_x2 =c("qt","eq")[1]
        , nest_x2 =  T
        , n1 = 10
        , n2 = 10
        , min_recs = 7
        , precision = 0
      )


      res = assign_strata(
       str_test
        ,dat_test
      )

      res$stratum

      summary(lm(height ~ dbh , data = res))
      summary(lm(height ~ dbh_cat , data = res))
      summary(lm(height ~ stratum , data = res))

      #example 2 numeric and numeric
      str_test1 = make_strata(
        dat_test
        , x1="height"
        , x2="dbh"
        , split_x1 =c("qt","eq")[1]
        , split_x2 =c("qt","eq")[1]
        , nest_x2 =  T
        , n1 = 10
        , n2 = 10
        , min_recs = 7
        , precision = 0
      )

      res1 = assign_strata(
        str_test1
        ,dat_test
      )

      #example 3 numeric and factor
      str_test2 = make_strata(
        dat_test
        , x1="height"
        , x2="dbh_cat"
        , split_x1 =c("qt","eq")[1]
        , split_x2 =c("qt","eq")[1]
        , nest_x2 =  T
        , n1 = 10
        , n2 = 10
        , type_x1 = "numeric"
        , type_x2 = "factor"
        , min_recs = 7
        , precision = 0
      )

      res2 = assign_strata(
        str_test2
        ,dat_test
      )



}
\author{
Jacob Strunk <jacob@some.domain>
}
