% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVEL_biomass.R
\name{NVEL_biomass}
\alias{NVEL_biomass}
\title{Predict biomass using National Volume Estimator Library (NVEL)}
\usage{
NVEL_biomass()
}
\arguments{
\item{dfTL}{tree list with a minimum of diamter and height columns}

\item{bioeq}{(optional) use single equation for all trees}

\item{bioeqNm}{(optional) column name in DFTL: provide region, forest, district for every tree in dfTL}

\item{spcdNm}{(required) column name in DFTL: USFS species code}

\item{dbhNm}{(required) column name in DFTL: dbh in inches}

\item{htNm}{(required) column name in DFTL: height in feet}

\item{clNm}{(optional) column name in DFTL: crown length in feet}

\item{merchDbNm}{(optional) column name in DFTL: upper db miminum for saw (in)}

\item{htPrd1Nm}{(optional) column name in DFTL: Height to the minimum top diameter inside bark for primary product (ft)}

\item{htPrd2Nm}{(optional) column name in DFTL: Height to the minimum top diameter inside bark for secondary product (ft)}

\item{dll_64}{path to 64bit dll}

\item{dll_32}{path to 64bit dll}

\item{load_dll}{T/F should dll be loaded (in case it is already loaded)}

\item{dll_func_vol}{name of volume prediction function call in NVEL .dll}

\item{dll_func_bioeq}{name of volume equation chooser function call in NVEL .dll}

\item{nclus}{number of cores to use}
}
\value{
a data.frame reformatted to include all of the original columns, any missing columns for the NVEL .dll, 15 columns of predicted volumes, and an error column
}
\description{
Predict biomass using National Volume Estimator Library (NVEL)
}
\details{
Predict biomass using National Volume Estimator Library (NVEL)

 see:
 https://www.fs.fed.us/forestmanagement/products/measurement/volume/nvel/index.php

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/19/2020 Implemented \cr
}
}
\examples{

        #look up volume equations
         library(RSForInvt)
         NVEL_voleq(region = 2, forest = "01",district = "01", spcd=951)
         NVEL_voleq(region = 2, forest = "01",district = "01", spcd=951)
         NVEL_voleq(region = 2, forest = "01",district = "01", spcd=rep(c(951,201),2))
         NVEL_voleq(dfTL=data.frame(region = 6, forest = "01",district = "01", spcd=rep(c(951,201),2)))

        #grab list of species
        if(!"dfSpp" \%in\% ls()){
          library(RSQLite)
          db0 = dbConnect(RSQLite::SQLite(), system.file("misc/NBEL/BiomassEqns.db", package="RSForInvt"))
          dfSpp = dbGetQuery(db0, paste("select * from tblspp"))
          dfCoeff = dbGetQuery(db0, paste("select * from BM_EQCoefs"))
          dbDisconnect(db0)
        }

        #build a fake tree list
        if("df_fake" \%in\% ls()){
          set.seed=111
          nfake=length(unique(dfCoeff$species_code))

          df_fake = data.frame(
            trid=1:(nfake)
            ,region = 6
            ,forest = "01"
            ,district = "01"
            ,dbh=10*abs(rnorm(nfake))
            ,ht=100*abs(rnorm(nfake))
            ,spcd = unique(dfCoeff$species_code)#'     sample(c("a","b","c","d") , nfake , T)
          )

        }

        #get volumes
        NVEL_volume( dfTL = df_fake )



}
\author{
Jacob Strunk <Jacob.strunk@usda.gov>
}
