% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_modeling.R
\name{OLS_modeling}
\alias{OLS_modeling}
\alias{reg_model}
\alias{lm_boot}
\alias{many_boots}
\alias{reg_multi}
\alias{mod_multi}
\alias{pred_multi}
\alias{multi_bs}
\alias{lm_summary}
\title{functions related to modeling}
\usage{
reg_model(
  dat_yx,
  reg_obj,
  rank = 1,
  rank_by = c("bic", "rsq", "rss", "adjr2", "cp"),
  debug = FALSE,
  ...
)

lm_boot(model, y = NA, data = NA, n_boot = 50)

many_boots(model, n_range, r_boots = 10, n_clus = 1, lm_boot, ...)

reg_multi(
  y_vars,
  data,
  form_y = c("y~.", "y~.*."),
  n_v_max = 3,
  n_best = 1,
  really_big = FALSE,
  n_clus = 1,
  ...
)

mod_multi(reg_multi_obj, data, ...)

pred_multi(
  mods,
  data,
  dat0,
  id_col,
  n_clus = 1,
  clus = NA,
  out_dir = NULL,
  se_fit = TRUE,
  return = FALSE,
  fix_outliers = T,
  outlier_fun = function(y0, yhat, yhat_se) {
     large = yhat > max(y0) * 1.5
    
    small = yhat < 0
     yhat[small] = 0
     yhat[large] = quantile(y0, 0.99) * 1.5
   
     yhat_se[large] = max(yhat_se[!large & !small])
     yhat_se[small] =
    max(yhat_se[!large & !small])
     data.frame(fit = yhat, se_fit = yhat_se)
 }
)

multi_bs(mods, n_boot = 50, n_clus, lm_boot)

lm_summary(x, data, resids = FALSE)
}
\arguments{
\item{dat_yx}{tbd}

\item{reg_obj}{tbd}

\item{rank}{tbd}

\item{rank_by}{tbd}

\item{debug}{tbd}

\item{...}{tbd
\cr\cr
\bold{pred_multi() parameters:}}

\item{model}{tbd}

\item{y}{tbd}

\item{data}{tbd}

\item{n_boot}{tbd}

\item{n_range}{tbd}

\item{r_boots}{tbd}

\item{n_clus}{tbd}

\item{lm_boot}{tbd
\cr\cr
\bold{lm_summary() parameters:}}

\item{y_vars}{tbd}

\item{form_y}{tbd}

\item{n_v_max}{tbd}

\item{n_best}{tbd}

\item{really_big}{tbd}

\item{reg_multi_obj}{tbd}

\item{mods}{tbd}

\item{dat0}{tbd}

\item{id_col}{tbd}

\item{clus}{tbd}

\item{out_dir}{tbd}

\item{se_fit}{tbd}

\item{return}{tbd}

\item{fix_outliers}{tbd}

\item{outlier_fun}{tbd
\cr\cr
\bold{multi_bs() parameters:}}

\item{x}{tbd}

\item{resids}{tbd}
}
\value{
reg_model:
lm_boot:
many_boots:
reg_multi:
mod_multi:
pred_multi:
multi_bs:
lm_summary:
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr

Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
}
}
\examples{

<Delete and Replace>




}
\seealso{
\code{\link{regsubsets}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
