% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs2.R
\name{pairs2}
\alias{pairs2}
\title{xy plot of variables in two dataframes}
\usage{
pairs2(
  x,
  y,
  str_max = 16,
  oneLine = F,
  mai_fig = c(0.3, 0.3, 0.3, 0.1),
  mai_edge = c(0, 0, 0, 0),
  title = NA,
  pdf_out = NA,
  pdf_ht = 8,
  pdf_wd = 11,
  xlab_fig = NA,
  ylab_fig = NA,
  combination = F,
  square = F,
  ...
)
}
\arguments{
\item{x}{first data.frame}

\item{y}{second data.frame}

\item{str_max}{number of characters per name to display}

\item{oneLine}{T/F add one to one line?}

\item{mai_edge}{"mai" parameters for composite figure}

\item{combination}{T/F combinations of x and y inputs or x[,1] vs y[,1] and x[,2] vs y[,2]}

\item{square}{T/F force graphs to be square (same x and y lims), only use if units match}

\item{...}{additional arguments to plot}

\item{tile}{optional figure title}
}
\value{
<Delete and Replace>
}
\description{
Prepare pair plots for two dataframes x,y: figure columns from x and rows from y. This results in fewer combinations than plot(xy_df)
}
\details{
<Delete and Replace>

\cr

Revision History
\tabular{ll}{
1.0 \tab 2015 Jan 29 Roxygen header prepared \cr
1.1 \tab 2022 March 03 added to RForInvt and added combination alternative \cr
}
}
\examples{

pairs2(mtcars,mtcars,combination=T)
pairs2(mtcars[,1:4],mtcars[5:8],combination=F)

}
\seealso{
\code{\link{plot}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
