% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_mod.R
\name{pander_mod}
\alias{pander_mod}
\title{update on the pander functions to do what I want to lm and simex objects including lists}
\usage{
pander_mod(x, ...)
}
\arguments{
\item{x}{inpute model or list of models (lm objects)}

\item{...}{additional arguments passed to methods}

\item{df0}{optional dataframe for validation (out of box validation)}

\item{return_df}{T/F should the function return a data.frame}
}
\value{
data.frame or pander table
}
\description{
simple output, and accepts a list of lm objects
}
\details{
Builds a compact one-row summary (response, n, residual std error, R^2, adjusted R^2,
 inverse design effect, and predictor list) for a fitted model and either returns it as a
 data.frame or prints it via \code{pander}. Methods are provided for \code{lm}, \code{simex},
 and a \code{list} of models (the list method row-binds the per-model summaries). When a
 validation data.frame \code{df0} is supplied, the statistics are computed out-of-sample.

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 Aug 22 \cr
}
}
\examples{
 fit1 = lm(mpg ~ wt + hp, data = mtcars)
 fit2 = lm(mpg ~ disp, data = mtcars)

 # one-row summary as a data.frame
 pander_mod(fit1, return_df = TRUE)

 # summarize a list of models (row-bound)
 pander_mod(list(a = fit1, b = fit2), return_df = TRUE)

}
\seealso{
\code{\link{pander}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
