% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_text.R
\name{search_text}
\alias{search_text}
\title{search for string in text files in a directory}
\usage{
search_text(
  match_string = "",
  directory = NA,
  file = NA,
  pattern = ".r",
  pattern_omit = c(NA, "dev_"),
  recursive = T,
  clus = NA
)
}
\arguments{
\item{match_string}{string to find within searched files}

\item{directory}{where to search files}

\item{file}{single file}

\item{pattern}{pattern to search for files}

\item{pattern_omit}{pattern to omit searched files}

\item{recursive}{look in subfolders?}

\item{clus}{number of cores, or a cluster object}
}
\value{
two column dataframe like:
 [1] file    matches
<0 rows> (or 0-length row.names)
}
\description{
search for string in text files in a directory
}
\details{
search for string in text files in a directory

\cr
Revision History
\tabular{ll}{
1.0 \tab 2015 June 19 Created \cr
}
}
\examples{
test=search_text(directory="//Dnrfsoly107/fr_data/forest_info_1/fris3/R/functions",pattern="[.]r",pattern_omit="dev_",match_string="require(plyr)")
print(test)
test=search_text(directory="//Dnrfsoly107/fr_data/forest_info_1/fris3/R/functions",pattern="[.]r",pattern_omit="dev_",match_string="library(plyr)")
print(test)
test=search_text(directory="//Dnrfsoly107/fr_data/forest_info_1/fris3/R/functions",pattern="[.]r",pattern_omit="dev_",match_string="data.table")
print(test)
test=search_text("c:\\\\temp\\\\process_gridmetrics",pattern="2015_Wed_Jul_29_17p01p50.*[.]bat$",match_string = "46896")
print(test)

}
\seealso{
\code{\link{makeCluster}}\cr \code{\link{list.files}}\cr  \code{\link{readLines}}\cr \code{\link{grepl}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
