% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yai_r2.R
\name{yai_r2}
\alias{yai_r2}
\title{Compute R2 values from yaImpute::yai() object or data frame returned by yaImpute::impute(...)}
\usage{
yai_r2(x, vars = NULL, ...)
}
\arguments{
\item{x}{yaImpute::yai() object or data frame returned by yaImpute::impute(...)}

\item{vars}{subset of variables}

\item{...}{additional arguments to impute() if yaImpute::yai() object is provided}
}
\value{
Data.frame with 1 column of R2 values
}
\description{
Compute R2 values from yaImpute::yai() object or data frame returned by yaImpute::impute(...)
}
\details{
This function is a repurposing of the yaImpute::rmsd() function to compute R.squared instead.


 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/27/2020 Created \cr
}
}
\examples{

 #fit model using one set of responses
 yaimod1 = yaImpute::yai( x = dat6[,c(vEnvironment1)	], y = dat6[,vBA1] , method = "gnn" )

 #impute a second set
 imp1 = impute(yaimod3,dat6[,vResp3])

 #compute R2
 yai_r2(imp1)


}
\seealso{
\code{\link{yaImpute::yai}}\cr \code{\link{yaImpute::impute}}\cr
}
\author{
Jacob Strunk <someone@somewhere.com>
}
